pyproject.toml
[tool.poetry]
name = "seaa2023_ect"
version = "0.1.0"
description = ""
authors = ["Santiago del Rey Juarez <santiago.del.rey@upc.edu>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
tensorflow-gpu = ">=2.3,<2.11"
opencv-python = "<4.8"
mlflow = ">=2.0, <2.2"
pysftp = "<0.3"
psutil = "<6.0"
numpy = "<1.24"
matplotlib = "<3.8"
tensorflow-addons = "^0.19.0"
seaborn = "^0.12.2"
statsmodels = "^0.13.5"
pingouin = "^0.5.3"
scikit-posthocs = "^0.7.0"
cliffs-delta = "^1.0.0"

[tool.poetry.group.dev.dependencies]
isort = "<5.12"
black = {extras = ["jupyter"], version = "<22.13"}
deptry = "<0.6"
jupyter = "<=1.0.0"
ipympl = "^0.9.3"

[tool.isort]
profile = "black"

[tool.black]
line-length=120

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
